src/app/features/animals/components/message/message.component.spec.ts
import { EventEmitter, SimpleChange, SimpleChanges } from '@angular/core';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { MatSnackBarModule } from '@angular/material/snack-bar';
import { MatSnackBarHarness } from '@angular/material/snack-bar/testing';
import { MatButtonHarness } from '@angular/material/button/testing';
import { HarnessLoader } from '@angular/cdk/testing';
import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';

import { createMagicalMock } from 'test.helper';
import { NoopAnimationsModule } from '@angular/platform-browser/animations';
import { MessageComponent } from './message.component';

describe('MessageComponent', () => {
  let component: MessageComponent;
  let fixture: ComponentFixture<MessageComponent>;
  let loader: HarnessLoader;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [MessageComponent],
      imports: [MatSnackBarModule, NoopAnimationsModule],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(MessageComponent);
    component = fixture.componentInstance;
    component.read = createMagicalMock(EventEmitter as EventEmitter<any>);
    loader = TestbedHarnessEnvironment.documentRootLoader(fixture);
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should display message', async () => {
    const changes: SimpleChanges = {
      message: new SimpleChange('', 'test', true),
    };

    component.ngOnChanges(changes);
    const snackBar = await loader.getHarness(MatSnackBarHarness);
    expect(await snackBar.getMessage()).toBe('test');

    const button = await loader.getHarness(MatButtonHarness);
    await button.click();
    expect(component.read.emit).toHaveBeenCalled();
  });
});
